+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [Unreleased](https://github.com/p-seven-v/illuminate-container-slim-bridge/compare/0.1.0...master)
 
+### Added
+- Added possibility to configure the container with service providers
+
 ## [0.1.0 (2020-08-16)](https://github.com/laravel/laravel/compare/v7.12.0...v7.25.0)
 
 ### Added